test: ssr test and drop hoisting do to refactor
