--- src/main/tests/hoisting.spec.tsx
@@ -1,117 +1,0 @@
-import { expect, test } from 'bun:test'
-import beautify from 'beautify'
-import { bElement, createStyles, joinStyles, type TemplateObject } from 'plaited'
-
-const render = (tpl: TemplateObject) => beautify(tpl.html.join(''), { format: 'html' })
-
-const componentStyles = createStyles({
-  nestedLabel: {
-    fontWeight: 'bold',
-  },
-  nestedComponent: {
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  slottedParagraph: {
-    color: 'rebeccapurple',
-  },
-  topComponent: {
-    display: 'block',
-  },
-  image: {
-    width: '100%',
-    aspectRatio: '16 / 9',
-  },
-})
-
-const NestedCustomElement = bElement({
-  tag: 'nested-component',
-  shadowDom: (
-    <>
-      <span {...componentStyles.nestedLabel}>inside nested template</span>
-      <slot name='nested' />
-    </>
-  ),
-})
-
-test('createTemplate: CustomElement hoisting its styles', () => {
-  const el = <NestedCustomElement />
-  expect({
-    content: render(el),
-    stylesheets: el.stylesheets,
-  }).toMatchSnapshot()
-})
-
-test('createTemplate: CustomElement with declarative shadow dom & hoist styles', () => {
-  const el = <NestedCustomElement {...componentStyles.nestedComponent} />
-  expect({
-    content: render(el),
-    stylesheets: el.stylesheets,
-  }).toMatchSnapshot()
-})
-
-test('createTemplate: CustomElement with styled slotted component', () => {
-  const el = (
-    <NestedCustomElement {...componentStyles.slottedParagraph}>
-      <p slot='nested'>slotted paragraph</p>
-    </NestedCustomElement>
-  )
-  expect({
-    content: render(el),
-    stylesheets: el.stylesheets,
-  }).toMatchSnapshot()
-})
-
-const TopCustomElement = bElement({
-  tag: 'top-component',
-  shadowDom: (
-    <NestedCustomElement {...componentStyles.slottedParagraph}>
-      <p slot='nested'>slotted paragraph</p>
-    </NestedCustomElement>
-  ),
-})
-
-test('createTemplate: CustomElement with declarative shadow dom and nested declarative shadow dom', () => {
-  const el = <TopCustomElement />
-  expect({
-    content: render(el),
-    stylesheets: el.stylesheets,
-  }).toMatchSnapshot()
-})
-
-test('createTemplate: CustomElement with declarative shadow dom and nested declarative shadow dom plus host styles', () => {
-  const el = <TopCustomElement {...componentStyles.topComponent} />
-  expect({
-    content: render(el),
-    stylesheets: el.stylesheets,
-  }).toMatchSnapshot()
-})
-
-test('createTemplate: CustomElement with declarative shadow dom and nested declarative shadow dom plus host styles and child', () => {
-  const el = (
-    <TopCustomElement {...componentStyles.topComponent}>
-      {/* biome-ignore lint/a11y/useAltText: Test fixture doesn't need alt text */}
-      <img {...componentStyles.image} />
-    </TopCustomElement>
-  )
-  expect({
-    content: render(el),
-    stylesheets: el.stylesheets,
-  }).toMatchSnapshot()
-})
-
-const hoistStyles = createStyles({
-  var1: {
-    width: '100%',
-  },
-  var2: {
-    width: '100%',
-  },
-  var3: {
-    color: 'blue',
-  },
-})
-
-test('ssr: Properly hoist and deduplicates multiple stylesheets on a single node', () => {
-  expect((<div {...joinStyles(hoistStyles.var1, hoistStyles.var2, hoistStyles.var3)} />).stylesheets.length).toBe(2)
-})
